+++ CHANGELOG.md
@@ -1,3 +1,9 @@
 ## 1.0.0
 
   - Initial release
+
+## 1.1.0
+
+  - Fixed the reading of the contents of `robots.txt`
+  - Fixed the parsing of rule fields to `Rule`s
+  - Added `example.dart`